+++ contracts/interfaces/IIndexPrice.sol
@@ -5,5 +5,5 @@ interface IIndexPrice {
     /// @notice Returns the index price of the token.
     /// @param index The interval represents twap interval.
     /// @return indexPrice Twap price with interval
-    function getIndexPrice(uint64 index) external view returns (uint256 indexPrice);
+    function getIndexPrice(uint256 index) external view returns (uint256 indexPrice);
 }
