+++ contracts/storage/FundingRateStorage.sol
@@ -9,7 +9,7 @@ abstract contract FundingRateStorage {
 
     address internal _markPriceOracleArg;
     address internal _indexPriceOracleArg;
-    uint64 internal _underlyingPriceIndex;
+    uint256 internal _underlyingPriceIndex;
     uint256 internal _fundingPeriod;
     mapping(address => uint256) internal _lastFundingIndexPrice;
